requirements.txt
flask==3.0.2
python-dotenv==1.0.1
requests==2.31.0
redis==5.0.1
flask-caching==2.1.0
flask-limiter==3.5.0
gunicorn==21.2.0 